Implement a much faster renderer that is capable of well over 100k sprites at >200FPS. Textures need to be reimplemented for this new renderer.

+++ src/graphics/buffers/IndexBuffer.cpp
@@ -1,12 +1,12 @@
 #include "IndexBuffer.h"
 
-IndexBuffer::IndexBuffer(GLushort *data, GLsizei count)
+IndexBuffer::IndexBuffer(GLuint *data, GLsizei count)
     : Count(count)
 {
 
     glGenBuffers(1, &BufferID);
     glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, BufferID);
-    glBufferData(GL_ELEMENT_ARRAY_BUFFER, count * sizeof(GLushort), data, GL_STATIC_DRAW);
+    glBufferData(GL_ELEMENT_ARRAY_BUFFER, count * sizeof(GLuint), data, GL_STATIC_DRAW);
     glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, 0);
 }
