+++ src/graphics/buffers/IndexBuffer.h
@@ -11,7 +11,7 @@ private:
     GLuint Count;
 
 public:
-    IndexBuffer(GLushort *data, GLsizei count);
+    IndexBuffer(GLuint *data, GLsizei count);
     ~IndexBuffer();
 
     void bind() const;
